fix(loan): revert unsafe post-submit editability, move regression test to loan disbursement suite
Drop allow_on_submit on maximum_loan_amount and the matching toggle_enable
in loan.js. Allowing the field to be edited after submit opens a wide,
unvalidated write path on a value that drives disbursal-limit and
security-shortfall checks for a loan that may already be disbursed.

Keep the get_disbursal_amount fallback to maximum_loan_amount when no
Loan Security Assignment exists, which is the actual fix for the 0.0
disbursal-limit bug.

Move the regression test to test_loan_disbursement.py since it asserts
on Loan Disbursement submission, and simplify it to set
maximum_loan_amount before submit instead of relying on a fabricated
post-submit db_set call that has no equivalent real code path.
