+++ lending/loan_management/doctype/loan/loan.json
@@ -357,7 +357,6 @@
    "read_only": 1
   },
   {
-   "allow_on_submit": 1,
    "depends_on": "eval:doc.is_secured_loan",
    "fieldname": "maximum_loan_amount",
    "fieldtype": "Currency",
